filter/tag file reorganization
