+++ src/features/catalog/components/CatalogCourse.tsx
@@ -5,10 +5,9 @@ import { IoAdd } from "react-icons/io5";
 
 import CourseBadge from "@/components/course/CourseBadge";
 import CourseLabel from "@/components/course/CourseLabel";
-import { findFiltersForCourse } from "@/components/course/utils";
+import Tag from "@/components/Tag";
 import { useCourseWorkspace } from "@/core/workspace/useCourseWorkspace";
-import { useCatalog } from "@/features/catalog/search/context/useCatalog";
-import Tag from "@/features/catalog/Tag";
+import { useCourseFilters } from "@/lib/stores/useFilterStore";
 import { APICourse } from "@/lib/types";
 
 interface CourseProps {
@@ -16,18 +15,8 @@ interface CourseProps {
 }
 
 const Course: React.FC<CourseProps> = ({ course }) => {
-  const { addCourseToToolbox } = useCourseWorkspace();
-  const { filters } = useCatalog();
-  const { getCourseCount } = useCourseWorkspace();
-
-  const attrFilters = findFiltersForCourse(
-    course.attr_list || [],
-    filters.attributes,
-  );
-  const semFilters = findFiltersForCourse(
-    course.sem_list || [],
-    filters.semesters,
-  );
+  const { addCourseToToolbox, getCourseCount } = useCourseWorkspace();
+  const { attrFilters, semFilters } = useCourseFilters(course);
 
   const [isOpen, setIsOpen] = useState(false);
   const toggleOpen = (e: React.MouseEvent<HTMLDivElement>) => {
